Created HStruct.h and moved a lot of non-LP structs to it from HighsLp.h
