+++ src/Highs.h
@@ -16,16 +16,10 @@
 
 #include <sstream>
 
-#include "lp_data/HighsInfo.h"
-#include "lp_data/HighsLp.h"
 #include "lp_data/HighsModelObject.h"
-#include "lp_data/HighsOptions.h"
 #include "lp_data/HighsRanging.h"
 #include "lp_data/HighsSolutionDebug.h"
-#include "lp_data/HighsStatus.h"
 #include "presolve/PresolveComponent.h"
-#include "util/HighsTimer.h"
-#include "util/HighsUtils.h"
 
 /**
  * @brief Class to set parameters and run HiGHS
