+++ src/lp_data/HighsModelObject.h
@@ -11,6 +11,7 @@
 #define LP_DATA_HIGHS_MODEL_OBJECT_H_
 
 #include "lp_data/HighsLp.h"
+#include "lp_data/HStruct.h"
 #include "lp_data/HighsOptions.h"
 #include "simplex/HFactor.h"
 #include "simplex/HMatrix.h"
